Separated node bindings to separate compilation units to reduce memory requirements
