+++ CMakeLists.txt
@@ -96,11 +96,34 @@ pybind11_add_module(${TARGET_NAME}
     src/DatatypeBindings.cpp
     src/DataQueueBindings.cpp
     src/pipeline/PipelineBindings.cpp
-    src/pipeline/NodeBindings.cpp
     src/pipeline/CommonBindings.cpp
     src/pipeline/AssetManagerBindings.cpp
     src/openvino/OpenVINOBindings.cpp
     src/log/LogBindings.cpp
+
+    src/pipeline/node/NodeBindings.cpp
+
+    src/pipeline/node/XLinkInBindings.cpp
+    src/pipeline/node/XLinkOutBindings.cpp
+    src/pipeline/node/ColorCameraBindings.cpp
+    src/pipeline/node/MonoCameraBindings.cpp
+    src/pipeline/node/StereoDepthBindings.cpp
+    src/pipeline/node/NeuralNetworkBindings.cpp
+    src/pipeline/node/VideoEncoderBindings.cpp
+    src/pipeline/node/ImageManipBindings.cpp
+    src/pipeline/node/SPIOutBindings.cpp
+    src/pipeline/node/SPIInBindings.cpp
+    src/pipeline/node/DetectionNetworkBindings.cpp
+    src/pipeline/node/SystemLoggerBindings.cpp
+    src/pipeline/node/ScriptBindings.cpp
+    src/pipeline/node/SpatialLocationCalculatorBindings.cpp
+    src/pipeline/node/SpatialDetectionNetworkBindings.cpp
+    src/pipeline/node/ObjectTrackerBindings.cpp
+    src/pipeline/node/IMUBindings.cpp
+    src/pipeline/node/EdgeDetectorBindings.cpp
+    src/pipeline/node/FeatureTrackerBindings.cpp
+    src/pipeline/node/AprilTagBindings.cpp
+    src/pipeline/node/DetectionParserBindings.cpp
 )
 
 if(WIN32)
