+++ src/pipeline/node/NodeBindings.hpp
@@ -7,6 +7,8 @@
 #include "depthai/pipeline/Node.hpp"
 
 struct NodeBindings : public dai::Node {
-    static void bind(pybind11::module& m, void* pCallstack);
+    static void addToCallstack(std::deque<StackFunction>& callstack);
     static std::vector<std::pair<py::handle, std::function<std::shared_ptr<dai::Node>(dai::Pipeline&, py::object class_)>>> getNodeCreateMap();
+ private:
+    static void bind(pybind11::module& m, void* pCallstack);
 };
